JIT Type parser for mobile
Pull Request resolved: https://github.com/pytorch/pytorch/pull/30391


A Type parser to parse the python string of a Type. For example,
"Tuple[str, Optional[float], Dict[str, List[Tensor]], int]".
Please refer to test_type_parser.cpp for the usage.

One of the use cases is in lite interpreter, types needs to be serialized (directly calling the python_str() of the Type) and deserialized (calling parseType(str)).

Differential Revision: [D18924268](https://our.internmc.facebook.com/intern/diff/D18924268/)
ghstack-source-id: 95422824